requirements.txt
Django==5.2.4
djangorestframework==3.15.2
python-dotenv==1.1.1
requests==2.32.4
groq==0.13.0
